Partially working property access assignment.
There is a major gotcha that I discovered. Binaryen does not support tuples inside of tuples. This means we don't yet support sub-structs or sub tuples on a language level.

The workaround is to flatten the structs to a linear representation.

That will be in the future though.

We have enough problems as it is.