Updated Functional Macro and Optional Chain Semantics (#316)
# Overview

This PR replaces the macro "unquote" operators $ and $@ with ~ and ~~
respectively. It also updates the `?.` optional chain operator to only
work on optional values.

As a result of this change, the `?.` is now defined as a functional
macro in std lib.

# Why the ~ change

$ and $@ acted as operators. So `$hi` would get turned into `($ hi)`.
This behavior prevented us from dynamically setting function names.
Because `~` is not an operator character (anymore), `~hi` is treated as
a uniform identifier by the parser. So the functional macro interpreter
can more predictably insert syntax without the parser from modifying the
syntax in unexpected ways.

Note: We also modify the syntax_macro function (fka quote) to unwrap
lists containing a single nested list. This is intentional and required.
Otherwise certain macros would be impossible due to how the parser
parses syntax templates.